pyproject.toml
[project]
name = "cardelephant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "beautifulsoup4>=4.14.2",
    "clip",
    "datasets>=4.2.0",
    "matplotlib>=3.4.0",
    "numpy>=1.21.0",
    "open-clip-torch>=3.2.0",
    "pandas>=2.3.2",
    "pillow>=8.3.0",
    "pre-commit>=4.3.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "seaborn>=0.11.0",
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "umap-learn>=0.5.9.post2",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git", rev = "dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1" }
